Improved the Makefile to handle devices and commands better.
Now all you have to do to add a device type is to create an
implementation of it in a file named "dev_*.c", and it will be
automatically built and included in the device type list with
a name taken from the file name.

Commands are similarly handled, but also with the added
improvement that you don't need a dummy "cmd_*.c" for commands
that are implemented elsewhere; so cmd_cpu.c has been removed.
Additionally, you can now create aliases for commands. See the
makefile for details. In the simple case that you want to add
a command with a single name, and implemented in its own file,
all you have to do is create the file and name it "cmd_*.c".

Also added stub implementations of two new commands "run" and
"step", mostly for testing the makefile.